feat + ref + fix + tests? see full commit message
Lexer is currently non working but more or less scaffolded. In the time
since the last commit, I've settled on implementing a version of Scheme
essentially equivalent to R5RS with some minor differences to simplify
implementation.

The main change will be the simplification of the
numbers. Numbers in standard Scheme are rather complex lexically so I've
decided to follow the simpler lexical spec as stated from this webpage:
https://amirkamil.github.io/project-scheme-parser/

I may potentially come back and implement this part to be inline with
the full spec but for the time being and in the interest of moving on
from writing lexers, I'm going with this.